3.5.0update

+++ src/c/src/zk_log.c
@@ -155,9 +155,8 @@ void log_message(log_callback_fn callback, ZooLogLevel curLevel,
 
 #ifndef THREADED
 
-    // pid_t is long on Solaris
     ofs = snprintf(buf, FORMAT_LOG_BUF_SIZE,
-                   "%s:%ld:%s@%s@%d: ", time, (long)pid,
+                   "%s:%d:%s@%s@%d: ", time, pid,
                    dbgLevelStr[curLevel], funcName, line);
 #else
 
@@ -168,7 +167,7 @@ void log_message(log_callback_fn callback, ZooLogLevel curLevel,
     #endif
 
     ofs = snprintf(buf, FORMAT_LOG_BUF_SIZE-1,
-                   "%s:%ld(0x%lx):%s@%s@%d: ", time, (long)pid, tid,
+                   "%s:%d(0x%lx):%s@%s@%d: ", time, pid, tid,
                    dbgLevelStr[curLevel], funcName, line);
 #endif
